clear ev flags when using an event timer
fixes issue where ping-ponging between event timers and epoll events was
avoiding going _back_ into the epoll even because the flags still matched.

this isn't purely correct, since it should be detecting if it was the same
type of event as before, but simply wiping all flags for timers works here.