add 'balloon' utility for soaking RAM
sometimes it's useful to waste RAM in a system quickly before doing a
benchmark or test. this is a really stupid utility that just allocates,
writes to, then holds on to RAM until killed.

for test systems with a lot of RAM it helps to use this to soak RAM rather
than pre-load hundreds of millions of items before every test