fix sock leak and improve pacing
missed removing old line of socket code which was doubling number of
sockets initialized (though only half were used).

use the real clock for pacing rather than relying solely on evtimer.

evtimer was very, very wobbly, especially at low times. now we attempt
to sleep for the right amount of time, and adjust up/down based on real
timings.

it's still a little wrong: every entry into the timer function means
we've already sent a blast of data, so the only option is to decide how
long to attempt to sleep before doing it again. this averages out close
to correct over a few runs, so keeping pacing down to tens of
milliseconds is probably necessary.