replace sprintf with memcpy + fast itoa from mc.
seems as fast as or faster than key_prealloc=1 in some tests.

mc-crusher's still heavy on main memory usage, but that seems to be
mostly caused by the kernel?

might be able to go from here to flat buffers. I want to use a
thread-based wbuf, and vivifying a c->wbuf only if necessary. EAGAIN on
write should only happen when we're flooding noreply commands.

even if that doesn't end up working out, cutting the writev is trivial.